renamed file from list_added_currencies... to added_currencies_list....
