+++ src/components/added_currencies_list.js
