+++ src/containers/added_currencies_list_container.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
 
-import AddedCurrenciesList from '../components/list_added_currencies';
+import AddedCurrenciesList from '../components/added_currencies_list';
 
 const AddedCurrenciesListContainerDiv = styled.div`
     grid-area: addedCurrencies;
